requirements.txt
aioftp==0.21.4
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
astropy==5.2.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
fonttools==4.39.3
frozenlist==1.3.3
gitdb==4.0.10
gitpython==3.1.31
idna==3.4
importlib-resources==5.12.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
multidict==6.0.4
numpy==1.24.2
packaging==23.1
pandas==2.0.0
parfive==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prettytable==3.9.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pyerfa==2.0.0.3
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
statsmodels==0.14.0
sunpy==4.1.5
tabulate==0.9.0
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
wcwidth==0.2.6
yarl==1.8.2
zipp==3.15.0
